+++ frontend/src/app/app-routing.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { NosotrosComponent } from './nosotros/nosotros.component';
-import { ContactanosComponent } from './contactanos/contactanos.component';
-import { PerfilComponent } from './perfil/perfil.component';
-import { CartaComponent } from './carta/carta.component';
-import { HomeComponent } from './home/home.component';
+import { NosotrosComponent } from './pages/nosotros/nosotros.component';
+import { ContactanosComponent } from './pages/contactanos/contactanos.component';
+import { PerfilComponent } from './auth/perfil/perfil.component';
+import { CartaComponent } from './pages/carta/carta.component';
+import { HomeComponent } from './pages/home/home.component';
 
 const routes: Routes = [
   {path: '', redirectTo:'/home', pathMatch:'full'},
