+++ frontend/src/app/app.module.ts
@@ -3,13 +3,13 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { NosotrosComponent } from './nosotros/nosotros.component';
-import { FooterComponent } from './footer/footer.component';
-import { ContactanosComponent } from './contactanos/contactanos.component';
-import { PerfilComponent } from './perfil/perfil.component';
-import { CartaComponent } from './carta/carta.component';
-import { HeaderComponent } from './header/header.component';
-import { HomeComponent } from './home/home.component';
+import { NosotrosComponent } from './pages/nosotros/nosotros.component';
+import { FooterComponent } from './shared/footer/footer.component';
+import { ContactanosComponent } from './pages/contactanos/contactanos.component';
+import { PerfilComponent } from './auth/perfil/perfil.component';
+import { CartaComponent } from './pages/carta/carta.component';
+import { HeaderComponent } from './shared/header/header.component';
+import { HomeComponent } from './pages/home/home.component';
 
 @NgModule({
   declarations: [
