+++ frontend/src/app/pages/carta/carta.component.spec.ts
