+++ frontend/src/app/pages/carta/carta.component.ts
