--- frontend/src/app/perfil/perfil.component.css
@@ -1,69 +1,0 @@
-
-.main-profile{
-    height: 100vh; 
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    background-color: #fdeedc;
-    padding-top: 15vh;
-  }
-  
-  .form-profile{
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-    padding: 1.5rem;
-    height: 60vh;
-    width: 35%;
-    margin-top: 3vh;
-    background-color: #ffd8a9;
-    box-shadow: 4px 4px 4px #666; 
-  }
-  .input-form{
-    border: 0 !important;
-    border-bottom: 2px solid #000000 !important;
-    background-color:#ffd8a9 !important;
-    margin: 1vh 0 1vh 0;
-  
-  }
-  
-  form .form-buttons .btn:active{
-    background-color: #e38b29 !important;
-  }
-  
-  .footer {
-    margin: 0;
-  }
-  
-  /* MEDIA QUERIES */
-  
-  /* TABLET */
-  
-  @media only screen and (max-width: 768px) {
-    .form-profile{
-      width: 60%;
-      height: 65vh;
-    }
-  }
-  
-  /* MOBILE */
-  
-  @media only screen and (max-width: 425px) {
-    /* PERFILES */
-    .form-profile{
-      width: 85%;
-      height: 75vh;
-      margin: 0;
-    }
-  
-    .form-buttons{
-      display: flex;
-      flex-direction: column;
-    }
-  
-    form .form-buttons .btn{
-      margin: 1vh !important;
-    }
-  }
