+++ frontend/src/app/shared/footer/footer.component.spec.ts
